requirements.txt
Django>=3.2,<4.0
elevenlabs
moviepy
Pillow
pysrt
pydub
numpy
aeneas
uvicorn
python-multipart
django-cors-headers==3.13.0
djangorestframework==3.14.0 